+++ src/hub/alerts/index.ts
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2022 The Pybricks Authors
+
+import { userProgramSize } from './UserProgramSize';
+
+export default {
+    userProgramSize,
+};